Remove KotlinModulePlatform
